Functionise combinatoric function
